+++ src/features/items/debugItem.ts
@@ -2,7 +2,7 @@ import { debugFunction } from "../../debug";
 
 // ModCallback.POST_USE_ITEM (3)
 // CollectibleTypeCustom.DEBUG
-export function useItemDebug(): boolean {
+export function postUseItemDebug(): boolean {
   debugFunction();
   return true; // Display the "use" animation
 }