+++ src/features/items/flipCustom.ts
@@ -137,7 +137,9 @@ function featureEnabled() {
 
 // ModCallback.POST_USE_ITEM (3)
 // CollectibleTypeCustom.FLIP_CUSTOM
-export function useItemFlipCustom(player: EntityPlayer): boolean | undefined {
+export function postUseItemFlipCustom(
+  player: EntityPlayer,
+): boolean | undefined {
   if (!config.flipCustom) {
     return undefined;
   }